forgot to move getThreds
